screw that for now. Continue with the re-construction
