+++ __tests__/fixtures/app/index.html
@@ -55,7 +55,7 @@
       <div class="starter-template">
         <h1>Bootstrap Template test for gulp-server-io</h1>
         <p class="lead">
-          Add something back, again
+          Add something back, again not array? 
           <br />
           Use this document as a way to quickly start any new project.
           <br>
