+++ src/index.js
@@ -1,2 +1,4 @@
-
-//const http = require('./lib/socket.js');
+/**
+ * The main server that wrap in the stream
+ */
+// module.exports = function(opt = {}) {};
